be/requirements.txt
flask
python-dotenv
mysql-connector-python
flask-cors